Server/requirements.txt
django
djangorestframework-simplejwt
djangorestframework
pillow
django-filter
django-cors-headers